requirements.txt
django~=2.2.7
django-suit
sorl-thumbnail
Pillow
django-admin-rangefilter
django-ckeditor
django-import-export
django-jet
psycopg2
psycopg2-binary
django-extensions
graphviz
djangorestframework==3.9.2
markdown
django-filter
django-rest-swagger
qrcode
pyqrcode
pypng
django-googledrive-storage
django-storages[google]
celery==4.2.0
django_celery_results
redis
django-redis
graphene-django
exponent_server_sdk
gunicorn
django_select2
django-cors-headers
elasticsearch-dsl
django-colorfield
social-auth-app-django
django-extensions
django-social-share
gunicorn
django-celery-results
celery[redis]
django-allauth
django-cookies-samesite
django-imagefit
django-qrcode
django_ace